+++ packages/utils/src/schema/retrieveSchema.ts
@@ -6,7 +6,9 @@ import {
   ADDITIONAL_PROPERTIES_KEY,
   ADDITIONAL_PROPERTY_FLAG,
   ALL_OF_KEY,
+  ANY_OF_KEY,
   DEPENDENCIES_KEY,
+  ONE_OF_KEY,
   REF_KEY,
 } from "../constants";
 import findSchemaDefinition, {
@@ -22,7 +24,7 @@ import {
   StrictRJSFSchema,
   ValidatorType,
 } from "../types";
-import getMatchingOption from "./getMatchingOption";
+import getFirstMatchingOption from "./getFirstMatchingOption";
 
 /** Resolves a conditional block (if/else/then) by removing the condition and merging the appropriate conditional branch
  * with the rest of the schema
@@ -205,6 +207,14 @@ export function stubExistingAdditionalProperties<
         );
       } else if ("type" in schema.additionalProperties!) {
         additionalProperties = { ...schema.additionalProperties };
+      } else if (
+        ANY_OF_KEY in schema.additionalProperties! ||
+        ONE_OF_KEY in schema.additionalProperties!
+      ) {
+        additionalProperties = {
+          type: "object",
+          ...schema.additionalProperties,
+        };
       } else {
         additionalProperties = { type: guessType(get(formData, [key])) };
       }
@@ -310,7 +320,7 @@ export function resolveDependencies<
   let resolvedSchema: S = remainingSchema as S;
   if (Array.isArray(resolvedSchema.oneOf)) {
     resolvedSchema = resolvedSchema.oneOf[
-      getMatchingOption<T, S, F>(
+      getFirstMatchingOption<T, S, F>(
         validator,
         formData,
         resolvedSchema.oneOf as S[],
@@ -319,7 +329,7 @@ export function resolveDependencies<
     ] as S;
   } else if (Array.isArray(resolvedSchema.anyOf)) {
     resolvedSchema = resolvedSchema.anyOf[
-      getMatchingOption<T, S, F>(
+      getFirstMatchingOption<T, S, F>(
         validator,
         formData,
         resolvedSchema.anyOf as S[],