+++ packages/utils/test/schema/index.ts
@@ -1,11 +1,13 @@
 import getDefaultFormStateTest from "./getDefaultFormStateTest";
 import getDisplayLabelTest from "./getDisplayLabelTest";
-import getMatchingOptionTest from "./getMatchingOptionTest";
+import getClosestMatchingOptionTest from "./getClosestMatchingOptionTest";
+import getFirstMatchingOptionTest from "./getFirstMatchingOptionTest";
 import isFilesArrayTest from "./isFilesArrayTest";
 import isMultiSelectTest from "./isMultiSelectTest";
 import isSelectTest from "./isSelectTest";
 import mergeValidationDataTest from "./mergeValidationDataTest";
 import retrieveSchemaTest from "./retrieveSchemaTest";
+import sanitizeDataForNewSchemaTest from "./sanitizeDataForNewSchemaTest";
 import toIdSchemaTest from "./toIdSchemaTest";
 import toPathSchemaTest from "./toPathSchemaTest";
 
@@ -14,12 +16,14 @@ export * from "./types";
 export {
   getDefaultFormStateTest,
   getDisplayLabelTest,
-  getMatchingOptionTest,
+  getClosestMatchingOptionTest,
+  getFirstMatchingOptionTest,
   isFilesArrayTest,
   isMultiSelectTest,
   isSelectTest,
   mergeValidationDataTest,
   retrieveSchemaTest,
+  sanitizeDataForNewSchemaTest,
   toIdSchemaTest,
   toPathSchemaTest,
 };