+++ packages/validator-ajv8/test/utilsTests/schema.test.ts
@@ -4,12 +4,14 @@ import Ajv2020 from "ajv/dist/2020";
 import {
   getDefaultFormStateTest,
   getDisplayLabelTest,
-  getMatchingOptionTest,
+  getClosestMatchingOptionTest,
+  getFirstMatchingOptionTest,
   isFilesArrayTest,
   isMultiSelectTest,
   isSelectTest,
   mergeValidationDataTest,
   retrieveSchemaTest,
+  sanitizeDataForNewSchemaTest,
   toIdSchemaTest,
   toPathSchemaTest,
 } from "@rjsf/utils/test/schema";
@@ -19,12 +21,14 @@ const testValidator = getTestValidator({});
 
 getDefaultFormStateTest(testValidator);
 getDisplayLabelTest(testValidator);
-getMatchingOptionTest(testValidator);
+getClosestMatchingOptionTest(testValidator);
+getFirstMatchingOptionTest(testValidator);
 isFilesArrayTest(testValidator);
 isMultiSelectTest(testValidator);
 isSelectTest(testValidator);
 mergeValidationDataTest(testValidator);
 retrieveSchemaTest(testValidator);
+sanitizeDataForNewSchemaTest(testValidator);
 toIdSchemaTest(testValidator);
 toPathSchemaTest(testValidator);
 
@@ -32,12 +36,14 @@ const testValidator2019 = getTestValidator({ AjvClass: Ajv2019 });
 
 getDefaultFormStateTest(testValidator2019);
 getDisplayLabelTest(testValidator2019);
-getMatchingOptionTest(testValidator2019);
+getClosestMatchingOptionTest(testValidator2019);
+getFirstMatchingOptionTest(testValidator2019);
 isFilesArrayTest(testValidator2019);
 isMultiSelectTest(testValidator2019);
 isSelectTest(testValidator2019);
 mergeValidationDataTest(testValidator2019);
 retrieveSchemaTest(testValidator2019);
+sanitizeDataForNewSchemaTest(testValidator2019);
 toIdSchemaTest(testValidator2019);
 toPathSchemaTest(testValidator2019);
 
@@ -45,11 +51,13 @@ const testValidator2020 = getTestValidator({ AjvClass: Ajv2020 });
 
 getDefaultFormStateTest(testValidator2020);
 getDisplayLabelTest(testValidator2020);
-getMatchingOptionTest(testValidator2020);
+getClosestMatchingOptionTest(testValidator2020);
+getFirstMatchingOptionTest(testValidator2020);
 isFilesArrayTest(testValidator2020);
 isMultiSelectTest(testValidator2020);
 isSelectTest(testValidator2020);
 mergeValidationDataTest(testValidator2020);
+retrieveSchemaTest(testValidator2020);
 retrieveSchemaTest(testValidator2020);
 toIdSchemaTest(testValidator2020);
 toPathSchemaTest(testValidator2020);